agentic-shared(move-credentials,mechanical_provable): move the provider credential contract into miles.rollout.agentic.credentials
PROVIDER_CREDENTIALS, forward_address, sandbox_key_supply and preflight_sdk
leave openenv_launch_common.py unchanged; the launcher imports them.
